Source Files
Took me a while but i finally got myself to release them. Full source files even with a few uncommitted changes if im not mistaken. Personally i think Open Source Software is the future and therefor i think making my project is one of the final steps of making it better. Im not sure if i will update it again or not but for now this is the latest commit.